DB Creation & CSV Importing Practice
In our endeavor to initiate this project, the first measure we undertook was to establish a database that would enable us to import and manage the data from the .CSV file. The .SQL file comprises of the necessary codes to construct the database, which will serve as a repository for all information derived from the .CSV file. For verification purposes, I have attached a screenshot demonstrating the accuracy of the database.