COVID-19 DB Creation Script & DB Design Diagram
Subsequently, we were provided with an incomplete script for creating a COVID-19 database. Our task was to utilize our comprehensive database design diagram created in Microsoft Visio to address the missing elements in the code. The final iteration of the database creation script can be found in the accompanying .SQL file. To facilitate a better understanding of the process, I have included the database design diagram, showcasing how the design was transformed into a functional database.