COVID-19 DB Data Implementation
We were presented with a script file containing the data to be integrated into the COVID-19 database that we had established. The database population script was comprehensive in nature, housing a significant amount of data. This presented us with the opportunity to understand the potential duration of the data implementation process based on the magnitude of information to be processed.