Manual Data Manipulation
We crafted a script for manual data entry, and were subsequently assigned with the responsibility of revising the title of one of the researchers following a promotion. Furthermore, we had to adjust the database to reflect the retirement of one of the researchers. To enhance the ease of tracking and understanding the various actions performed, I emphasized the inclusion of clear and concise comments throughout the script.