Creating SQL Queries for Healthcare Database Analysis
The purpose of this project is to apply knowledge of data query language (DQL) to retrieve specific sets of data from the COVID-19 database. As part of this project, you will be expected to create single-table query statements using SQL to extract meaningful information from the database.

The outcome of this project is to retrieve specific sets of data from the COVID-19 database, using SQL DQL statements. The results of your queries will provide valuable insights into the spread and impact of COVID-19, and will contribute to our understanding of the pandemic.